requirements.txt
flask
gunicorn
pytesseract
pillow